Fix int/float mixups in stat collection

+++ weasyl/polecat.py
@@ -84,7 +84,7 @@ class WeasylSite(Site):
 
     def readRequestCount(self):
         "Return and reset the request count, request error percentage, and maximum active client count."
-        errorPercentage = 0
+        errorPercentage = 0.0
         if self.requestCount:
             errorPercentage = self.errorCount / self.requestCount * 100
         ret = self.requestCount, errorPercentage, self.mostActiveClients
@@ -218,7 +218,7 @@ class FetchRequestBreakdownStats(amp.Command):
         ('totalTimeInSQL', amp.Float(optional=True)),
         ('totalTimeInMemcached', amp.Float(optional=True)),
         ('totalTimeInPython', amp.Float(optional=True)),
-        ('queries', amp.ListOf(amp.Float(), optional=True)),
+        ('queries', amp.ListOf(amp.Integer(), optional=True)),
     ]
